Support DS5 devices in USB2 configuration Add helper class to query the actual device type Update device id when connected in USB2 mode

+++ common/rendering.h
@@ -269,7 +269,7 @@ namespace rs2
                 H *= scale;
             }
 
-            return{ x + floor(w - W) / 2, y + floor(h - H) / 2, W, H };
+            return{ x + float(floor(w - W) / 2), y + float(floor(h - H) / 2), W, H };
         }
 
         rect scale(float factor) const
